+++ src/main/java/dabang/star/cafe/api/request/RegisterParam.java
@@ -1,0 +1,28 @@
+package dabang.star.cafe.api.request;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonRootName;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+import java.time.LocalDate;
+
+@Data
+@JsonRootName("user")
+@NoArgsConstructor
+public class RegisterParam {
+    @NotBlank(message = "can't be empty")
+    @Email(message = "should be an email")
+    private String email;
+    @NotBlank(message = "can't be empty")
+    private String password;
+    @NotBlank(message = "can't be empty")
+    private String nickname;
+    @NotBlank(message = "can't be empty")
+    private String tel;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
+    private LocalDate birthday;
+}
+
