--- src/main/java/dabang/star/cafe/application/UserQueryService.java
@@ -1,19 +1,0 @@
-package dabang.star.cafe.application;
-
-import dabang.star.cafe.application.data.UserData;
-import dabang.star.cafe.infrastructure.mybatis.readservice.UserReadService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class UserQueryService {
-
-    private final UserReadService userReadService;
-
-    public Optional<UserData> findById(String id) {
-        return Optional.ofNullable(userReadService.findById(id));
-    }
-}
